requirements.txt
langchain==0.0.333
pypdf==3.17.0
chromadb==0.4.16
sentence-transformers==2.2.2
python-dotenv==1.0.0
esprima==4.0.1
atlassian-python-api==3.41.3
beautifulsoup4==4.12.2
pytesseract==0.3.10
reportlab==4.0.7
svglib==1.5.1
rlPyCairo==0.3.0
pycairo==1.25.1
pdf2image==1.16.3
docx2txt==0.8
slack-sdk==3.23.0
schedule==1.2.1
GitPython==3.1.40
python-dotenv==1.0.0